selftests/bpf: Fix ASSERT_EQ/NEQ macro portability
Macros ASSERT_EQ() and ASSERT_NEQ() are sometimes used to test pointers
against pointer constants, but their built-in casting leads to compile
errors on 32-bit systems:

In file included from .../selftests/bpf/prog_tests/btf_sysfs.c:4:
.../selftests/bpf/prog_tests/btf_sysfs.c: In function ‘test_btf_mmap_sysfs’:
./test_progs.h:256:23: error: cast from pointer to integer of different size [-Werror=pointer-to-int-cast]
  256 |               (name), (long long)(___act), (long long)(___exp));   \
      |                       ^

Update ASSERT_EQ() and ASSERT_NEQ() with a transitional 'long' cast to work
portably with pointers.

+++ tools/testing/selftests/bpf/test_progs.h
@@ -253,7 +253,7 @@ void hexdump(const char *prefix, const void *buf, size_t len);
 	bool ___ok = ___act == ___exp;					\
 	CHECK(!___ok, (name),						\
 	      "unexpected %s: actual %lld != expected %lld\n",		\
-	      (name), (long long)(___act), (long long)(___exp));	\
+	      (name), (long long)(long)(___act), (long long)(long)(___exp));   \
 	___ok;								\
 })
 
@@ -264,7 +264,7 @@ void hexdump(const char *prefix, const void *buf, size_t len);
 	bool ___ok = ___act != ___exp;					\
 	CHECK(!___ok, (name),						\
 	      "unexpected %s: actual %lld == expected %lld\n",		\
-	      (name), (long long)(___act), (long long)(___exp));	\
+	      (name), (long long)(long)(___act), (long long)(long)(___exp));   \
 	___ok;								\
 })
